python-lib/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kiassist-utils"
version = "0.1.0"
description = "Utility library for KiAssist - KiCAD AI Assistant"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pywebview>=5.0",
    "requests>=2.31.0",
    "keyring>=24.0.0",
]
authors = [
    {name = "KiAssist Contributors"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[project.urls]
Homepage = "https://github.com/mylonics/NewKiAssist"
Repository = "https://github.com/mylonics/NewKiAssist"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pyinstaller>=6.0.0",
]

[project.scripts]
kiassist = "kiassist_utils.main:main"
